+++ spec/spec_helper.rb
@@ -19,6 +19,9 @@ require 'simplecov' if ENV['COVERAGE']
 require 'fast_ignore'
 
 require_relative 'support/temp_dir_helper'
+require_relative 'support/stub_env_helper'
+require_relative 'support/stub_file_helper'
+require_relative 'support/stub_global_gitignore_helper'
 require_relative 'support/matchers'
 
 RSpec.configure do |config|
@@ -29,6 +32,10 @@ RSpec.configure do |config|
 
   config.mock_with :rspec do |c|
     c.verify_partial_doubles = true
+  end
+
+  config.before do
+    stub_blank_global_config
   end
 
   config.example_status_persistence_file_path = '.rspec_status'